Import hs-test-framework-quickcheck2-0.2.10 as wip/hs-test-framework-quickcheck2.
Allows QuickCheck2 properties to be used with the test-framework
package.

Due to technical limitations of the QuickCheck2 package, it's not
possible to correlate arguments that caused the failure with the
particular Test that caused them. This means that the arguments to
your properties that lead to failure will be printed at
non-deterministic positions in the stdout stream. Sorry.