Mainly an upgrade of gcompris version from a very old one to current 12.01.
It is found that this gcompris version does not work well with gstreamer.
Instead SDL mixer option is used which works pretty well.

Accordingly, --enable-sdlmixer added by default. As a fallout of this MESSAGE
file that asked to install gst-plugins deleted.

Buildlink added for sdl-mixer, py-cairo, librsvg, libxml2 due to new
dependencies.

configure script has made gnet a default option. So instead of enable-gnet
disable-gnet code is added to Makefile depending on pkg option.

Patches from games/gcompris are still applicable and are taken in with due
modifications.

PKGREVISION reset to 1.