Reintroduce local mythes distfile, unbreak build.
Part of the build system expects this distfile to be present, even if it
is not used to produce the final binary.
Just use the tarball again for now instead of spending weeks trying to
understand and patch the mess. This brain damaged build code will
eventually be rewritten to a saner gmake based system upstream anyway.