This is a big upgrade of the node pkg. tl;dr 0.4.7 -> 0.7.9 (RC0) tl;dr * fix do-configure tl;dr * lockf, flock, and shlock mess
---

It takes us to a pre-release version instead of a very old verion. :)

Many of our patchss have been accepted upstream by v8, libuv, and node.

uv/src/unix/netbsd.c DEFINITELY needs some extra work as it's just hacked.

There is an undelcared dependency on flock, which shlock could replace, but
it needs to be handled more cleanly.
(I hoped mine from freebsd and made a symlink)

Not sure if libexecinfo is still needed, but I kept it in there.  I was able
to build without it locally.

this is obviously the biggest hack:
do-configure:
        cd ${WRKSRC} && export GYP_DEFINES="OS=netbsd"; ./configure ${CONFIGURE_ARGS}

but it got me working on netbsd so I'll take it.