Import plinkseq-0.08 as wip/plinkseq.
PLINK/SEQ is an open-source C/C++ library for working with human genetic
variation data. The specific focus is to provide a platform for analytic tool
development for variation data from large-scale resequencing and genotyping
projects, particularly whole-exome and whole-genome studies. It is independent
of (but designed to be complementary to) the existing PLINK package.