Import py27-fatiando-0.0.1 as wip/py-fatiando.
Fatiando a Terra provides command-line tools and an API for a variety of
geophysical modeling applications. It makes heavy use of Numpy for linear
algebra operations, provides a high level interface to visualise your data
and results using Matplotlib for 2D and Mayavi2 for 3D. Optimized code is
written in Fortran and C and wrapped to Python using Numpy's f2py.