Import eet-1.7.4 as wip/eet.
EET is a tiny library designed to write an arbitrary set of chunks of
data to a file and optionally compress each chunk (very much like a
zip file) and allow fast random-access reading of the file later on.
It does not do zip as a zip itself has more complexity than is needed,
and it was much simpler to implement this once here.

EET is extremely fast, small and simple.  EET files can be very small
and highly compressed, making them very optimal for just sending
across the Internet without having to archive, compress or decompress
and install them.  They allow for lightning-fast random-access reads
once created, making them perfect for storing data that is written
once (or rarely) and read many times, but the program does not want to
have to read it all in at once.

It also can encode and decode data structures in memory, as well as
image data for saving to EET files or sending across the network to
other machines, or just writing to arbitrary files on the system.  All
data is encoded in a platform independent way and can be written and
read by any architecture.