Import amaya-11.4.4 as wip/amaya.
Amaya is a complete web browsing and authoring environment and
comes equipped with a WYSIWYG style of interface, similar to that
of the most popular commercial browsers.  With such an interface,
users do not need to know the HTML or CSS languages. Amaya does
not support Javascript and has an incomplete CSS implementation.
Amaya supports XHTML, MathML, SVG, and RDF.