Import lldpd-0.7.2 as wip/lldpd.
lldpd is an ISC-licensed implementation of LLDP for various Unixes.
It also supports some proprietary protocols. LLDP is an industry standard
protocol designed to supplant proprietary Link-Layer protocols such as
EDP or CDP. The goal of LLDP is to provide an inter-vendor compatible
mechanism to deliver Link-Layer notifications to adjacent network devices.