Import postgresql91-pgpool2-3.2.4 as wip/pgpool2.
pgpool-II is a middleware that works between PostgreSQL
servers and a PostgreSQL database client. It talks
PostgreSQL's backend and frontend protocol, and relays
a connection between them. Therefore, a database application
(frontend) thinks that pgpool-II is the actual PostgreSQL
server, and the server (backend) sees pgpool-II as one
of its clients.