Import unzip-6.10b as wip/unzip.
Unzip will list, test, or extract files from a ZIP archive, commonly
found on MS-DOS systems.  The default behavior (with no options) is to
extract into the current directory (and subdirectories below it) all
files from the specified ZIP archive. Unzip is compatible with
archives created by PKWARE's PKZIP, but in many cases the program
options or default behaviors differ.

Zipinfo lists technical information about files in a ZIP archive, most
commonly found on MS-DOS systems.  Such information includes file access
permissions, encryption status, type of compression, version and operating
system or file system of compressing program, and the like.

Funzip acts as a filter; that is, it assumes that a ZIP archive is
being piped into standard input, and it extracts the first member from
the archive to stdout.  If there is an argument, then the input comes
from the specified file instead of from stdin.

Unzipsfx is a modified version of unzip designed to be prepended to
existing ZIP archives in order to form self-extracting archives.
Instead of taking its first non-flag argument to be the zipfile(s) to
be extracted, unzipsfx seeks itself under the name by which it was
invoked and tests or extracts the contents of the appended archive.