Import simulavr-1.0.0 as wip/simulavr.
The Simulavr program is a simulator for the Atmel AVR family of
microcontrollers. Simulavr can be used either standalone or as a
remote target for gdb. When used in gdbserver mode, the simulator
is used as a backend so that gdb can be used as a source level
debugger for AVR programs.

Simulavr was written by Theodore A. Roth

(This is BROKEN now.)