Import libconfuse-2.7 as wip/libconfuse.
This is libConfuse, a library for parsing configuration files. It was
called libcfg before, but was changed to not confuse with other
similar libraries.

See the INSTALL file for installation instructions. By default, only a
static library is built. Specify --enable-shared to the configure
script to also build a shared library.

With libtool 1.4.2 (and/or automake 1.6.3) there is a problem building
the sources. The libtool script expects the environment variable SED
to point to a usable sed binary, but this variable does not get
set. You'll have to do a "export SED=sed" (if using bash) or "setenv
SED sed" (if using csh) before running make.

Documentation can be built by running doxygen in the top-level
directory. It is also available at http://www.nongnu.org/confuse/manual/.

There is a project page at http://sv.nongnu.org/projects/confuse.
Report bugs to <confuse-devel@nongnu.org>