Import llvm-rtlib-3.3 as wip/llvm-rtlib.
The compiler-rt project is a simple library that provides an implementation
of the low-level target-specific hooks required by code generation and other
runtime components.

For example, when compiling for a 32-bit target, converting a double to
a 64-bit unsigned integer is compiling into a runtime call to
the "__fixunsdfdi" function.

The compiler-rt library provides optimized implementations of this and other
low-level routines.