Update to 0.42:
This version is in sync with MAME/MESS 0.152 (released earlier
today) and contains a number of major performance improvements,
especially for the loading of cached data. Despite the performance
improvements, we also have a few important stability fixes this
time, and Qt 5 support has been greatly improved as well (we still
recommend Qt 4.8, though). Please see the change-log below for more
details!