Import py27-nimfa-1.0 as wip/py-nimfa.
Nimfa is a Python scripting library which includes a number of published matrix
factorization algorithms, initialization methods, quality and performance
measures and facilitates the combination of these to produce new strategies.
The library represents a unified and efficient interface to matrix factorization
algorithms and methods.

The nimfa library works with numpy dense matrices and scipy sparse matrices
(where this is possible to save on space). The library has support for multiple
runs of the algorithms which can be used for some quality measures. By setting
runtime specific options tracking the residuals error within one (or more) run
or tracking fitted factorization model is possible.Extensive documentation with
working examples which demonstrate real applications, commonly used benchmark
data and visualization methods are provided to help with the interpretation and
comprehension of the results.