Import py27-drawnow-0.44 as wip/py-drawnow.
Matlab has a great feature where it allows you to update a figure. You can
simply call drawnow and have your figure update. This is nice if you're running
a simulation and want to see the results every iteration. It'd sure be nice if
Python/matplotlib had a similar feature to update the plot each iteration.
This package adds that functionality to matplotlib.